requirements.txt
aiofiles
dill==0.3.5.1
dnspython==2.2.1
database-connect==0.1.66
evidently==0.7.8
Flask==2.0.3
httptools==0.5.0
imblearn==0.0
jinja2==3.1.2
matplotlib
mypy-boto3-s3==1.24.76
Werkzeug==2.0.3
pip-chill==1.0.1
pymongo==4.2.0
python-dotenv==0.21.0
python-multipart==0.0.6
seaborn==0.12.2
types-s3transfer==0.6.0.post4
uvicorn==0.18.3
watchfiles==0.17.0
websockets==10.3
wincertstore==0.2
xgboost==1.6.2
neuro-mf==0.0.5
boto3==1.26.105
fastapi
nltk
numpy
pandas
setuptools
scikit-learn

-e .